Continue implement /sshkeys

+++ api/sshkeys.py
@@ -1,5 +1,5 @@
 from flask import request
-from models import RoleEnum, SSHKey, sshkey_schema
+from models import RoleEnum, SSHKey, sshkey_schema, sshkeys_schema
 from app import db, oidc
 from utils import oidc_require_role
 from werkzeug.exceptions import NotFound, BadRequest, Forbidden
@@ -18,20 +18,18 @@ def search(offset, limit):
     if not results:
         raise NotFound(description="No sshkeys have been found.")
 
-    res = []
-    for result in results:
-        res.append(result.public_key)
-
-    return res
+    return sshkeys_schema.dump(results).data
 
 
 # /POST /sshkeys
 @oidc_require_role(min_role=RoleEnum.user)
 def post(user):
-    sshkey_data = request.data
-    print(sshkey_data)
+    sshkey_data = request.get_json()
+    data = sshkey_schema.load(sshkey_data).data
 
-    sshkey = SSHKey(public_key=sshkey_data, user_id=user.id)
+    public_key = data["public_key"]
+
+    sshkey = SSHKey(public_key=public_key, user_id=user.id)
     db.session.add(sshkey)
     db.session.commit()
 
@@ -42,7 +40,8 @@ def post(user):
 
 # /DELETE /sshkeys/{skId}
 @oidc_require_role(min_role=RoleEnum.user)
-def delete(sshkey_id):
+def delete(sshkey_id, user):
+    # TODO: capsule owners should be able to delete sshkey
     try:
         sshkey = SSHKey.query.get(sshkey_id)
     except:
@@ -50,6 +49,9 @@ def delete(sshkey_id):
 
     if sshkey is None:
         raise NotFound(description=f"The requested sshkey '{sshkey_id}' has not been found.")
+
+    if (user.id is not sshkey.user_id) and (user.role < RoleEnum.admin):
+        raise Forbidden
 
     db.session.delete(sshkey)
     db.session.commit()
